+++ packages/extensions/core/src/lib/configuration/index.ts
@@ -1,6 +1,7 @@
 export * from "./configuration-loader";
 export * from "./autorest-raw-configuration";
-export * from "./configuration-view";
+export * from "./autorest-configuration";
+export * from "./autorest-context";
 export * from "./message-emitter";
 export * from "./directive";
 export * from "./configuration-file-resolver";
